requirements.txt
requests
langchain-community
langchain-core
pandas
fastapi
uvicorn[standard]
langchain
langchain-openai
wechatpy
python-dotenv
lxml
cryptography
openpyxl
dashscope
pdfplumber
psutil
pydantic
pytest
pytest-asyncio